Exposing port 80 on Elastic Beanstalk
In order for our container to be accessible to the outside world when
running on Elastic Beanstalk we need to tell it to expose port 80. That
is what this commit does.

Note: this has no effect on our local machine, this instruction is only
required by Elastic Beanstalk. To expose ports locally we need to use
`-p` still or `ports` inside our **docker-compose.yml** file.